libsmi: drop generated files, adjust gitignore
Some of the manually-patched changes re-committed back into SnmpB git.

Those Makefile.in's can be recreated by autoconf -- if you have the
automake toolchain installed!

If not, and that's an issue (like e.g. on windows): the libsmi release
tarball contains these files, and we carry the incremental patches. So
you can *still* re-create the Makefile.in's, just by using CMake (which
will download, unpack, and patch automatically).

With that, source-controlling these *here* seems like an unnecessary burden.